Move automaticupgrades packages in `lib/automaticupgrades`
