+++ integrations/kube-agent-updater/pkg/controller/deployment.go
@@ -30,7 +30,7 @@ import (
 	kclient "sigs.k8s.io/controller-runtime/pkg/client"
 	ctrllog "sigs.k8s.io/controller-runtime/pkg/log"
 
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/version"
+	"github.com/gravitational/teleport/lib/automaticupgrades/version"
 )
 
 // DeploymentVersionUpdater Reconciles a podSpec by changing its image
