+++ lib/automaticupgrades/maintenance/basichttp.go
@@ -27,9 +27,9 @@ import (
 	"github.com/gravitational/trace"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/basichttp"
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/cache"
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/constants"
+	"github.com/gravitational/teleport/lib/automaticupgrades/basichttp"
+	"github.com/gravitational/teleport/lib/automaticupgrades/cache"
+	"github.com/gravitational/teleport/lib/automaticupgrades/constants"
 )
 
 // basicHTTPMaintenanceClient retrieves whether the target version represents a
