+++ lib/automaticupgrades/version/basichttp.go
@@ -26,9 +26,9 @@ import (
 
 	"github.com/gravitational/trace"
 
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/basichttp"
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/cache"
-	"github.com/gravitational/teleport/integrations/kube-agent-updater/pkg/constants"
+	"github.com/gravitational/teleport/lib/automaticupgrades/basichttp"
+	"github.com/gravitational/teleport/lib/automaticupgrades/cache"
+	"github.com/gravitational/teleport/lib/automaticupgrades/constants"
 )
 
 // basicHTTPVersionClient retrieves the version from an HTTP endpoint
